fix: undefined adaptProduct prop, remove useless adapter, rename /next-pages
fix-details: undefined discount in adaptProduct was triggering a JSON unserialized error in productId page, also fix mismatch product returned and types in SSR of ProductID page
