+++ frontend/src/adapters/index.ts
@@ -1,4 +1,3 @@
 export * from './adaptAllProductCards.adapter'
-export * from './adaptAllProducts.adapter'
 export * from './adaptProduct.adapter'
 export * from './adaptUniqueProductValues.adapter'
