+++ frontend/src/components/layout/Layout.tsx
@@ -6,11 +6,14 @@ import { Footer } from './Footer'
 import { footerLinks, headerLinks } from './refs'
 import { isString, setUpperCase } from '@/utils'
 
-import type { RootLayout } from '_app-types'
+export interface RootLayout {
+  children: React.ReactNode
+  title: string
+  section?: string
+}
 
 export const Layout: React.FC<RootLayout> = ({ children, title, section }) => {
   const isTitleWithSection = isString(section) ? `${title} | ${setUpperCase(section as string)}` : `${title}`
-
   return (
     <>
       <Head>
