+++ frontend/src/next-pages.ts
@@ -1,17 +1,13 @@
 import type { AppProps } from 'next/app'
-import type { ReactElement, ReactNode } from 'react'
 import type { NextPage } from 'next'
+import type { ReactElement, ReactNode } from 'react'
+
+export type NextPageResult = ReactElement | null
 
 export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
   getLayout?: (page: ReactElement, pageProps: P) => ReactNode
 }
 
-type AppPropsWithLayout = AppProps & {
+export type AppPropsWithLayout = AppProps & {
   Component: NextPageWithLayout
-}
-
-export interface RootLayout {
-  children: ReactNode
-  title: string
-  section?: string
 }
