+++ frontend/src/pages/_app.tsx
@@ -1,8 +1,6 @@
 import { CartProvider } from '@/contexts/cartContext'
-
 import '@/styles/globals.css'
-
-import type { AppPropsWithLayout } from '_app-types'
+import type { AppPropsWithLayout } from '@/next-pages'
 
 export default function App ({ Component, pageProps }: AppPropsWithLayout): React.ReactNode {
   const getLayout = Component.getLayout ?? ((page) => page)
