+++ frontend/src/pages/user/[session].tsx
@@ -4,7 +4,7 @@ import { Layout } from '@/components/layout'
 import { LoginForm, RegisterForm } from '@/modules/user/components'
 import { sessions, SESSIONS } from '@/modules/user/refs'
 
-import type { NextPageWithLayout } from '_app-types'
+import type { NextPageWithLayout } from '@/next-pages'
 import type { GetStaticPathsResult, GetStaticPropsContext, GetStaticPropsResult } from 'next'
 import type { Params } from 'next/dist/shared/lib/router/utils/route-matcher'
