+++ frontend/src/services/fetchProductId.service.ts
@@ -1,12 +1,9 @@
 import { adaptProduct } from "@/adapters"
 import { Product, ProductIdResponse } from "@/models"
 
-const ITEMS_DISPLAYED = 12
-
 export const fetchProductId = async (id: string): Promise<Product> => {
   const url = `https://e-commerce-store-api.onrender.com/api/v1/products/${id}`
   const res = await fetch(url)
   const data: ProductIdResponse = await res.json()
-
   return adaptProduct(data)
 }
